@@ -10,7 +10,7 @@ dynamic = [
 ]
 description = "Easy configurable workers based on asyncio"
 license = "Apache-2.0"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 authors = [
     { name = "Alexander Malev", email = "malev@somedev.ru" },
 ]
@@ -24,7 +24,6 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
@@ -120,7 +119,6 @@ cov = [
 
 [[tool.hatch.envs.all.matrix]]
 python = [
-    "3.7",
     "3.8",
     "3.9",
     "3.10",
@@ -159,7 +157,6 @@ exclude_lines = [
 [tool.black]
 line-length = 120
 target-version = [
-    "py37",
     "py38",
     "py39",
     "py310",
